Add Whatif improvement

+++ Public/Remove-InfobloxFixedAddress.ps1
@@ -33,7 +33,9 @@
         if ($Output) {
             Write-Verbose -Message "Remove-InfobloxFixedAddress - Removed $($Mac.ipv4addr) with mac address $($Mac.mac) / $Output"
         } else {
-            Write-Warning -Message "Remove-InfobloxFixedAddress - Failed to remove $($Mac.ipv4addr) with mac address $($Mac.mac), error: $varWarning"
+            if (-not $WhatIfPreference) {
+                Write-Warning -Message "Remove-InfobloxFixedAddress - Failed to remove $($Mac.ipv4addr) with mac address $($Mac.mac), error: $varWarning"
+            }
         }
     }
 }
